requirements.txt
python-docx
httpx
pymupdf
ocrmypdf
PyPDF2
google-genai
langchain_google_genai